i #4 Add communication network parser and viz
The vignette was renamed to showcase various networks.

The mbox_downloader.sh was provided by Wolfgang. In a future
issue, this should be integrated into R code for full automation.

R/parsers.R have 2 new functions to use Perceval to parse mbox,
and to use the parsed mbox into an igraph object were added.
The vignette  uses the igraph object to create a
interactive visualization of the data.