i #12 Add MPL 2.0 License
As of this commit, the code only Imports the
following packages:

 * data.table (MPL 2.0)
 * stringi (BSD-3-clause)
 * jsonlite (Apache 2.0)

Moreover, the use of Perceval (GPL 3.0) is done by command line,
and can be replaced by any other commit log or mailing list log parsers,
as will be done in future commits. This, to the best of my knowledge,
as I am not a lawyer, complies with GPL based on:

https://www.gnu.org/licenses/gpl-faq.html#MereAggregation

and other open source packages like XGBoost (i #1338), in respect
to clarify the  codebase does not depend on a particular GPL project
to function.

Author roles were also rectified.