i #23 adds lang identifier and file filter
Adds a suggestion for the package gh to use
on this Vignette. The provided endpoint will
obtain the language percent of a project given
it's URL on Github. A language must be chosen
in order to use Depends.

A rudimentary string matching filter was added to
filter out file_paths which contain `test` and
`__init__.py` on the vignette only.