i #41 Creates conf file format
The package now suggests the use of yaml to create
config files. The dependency is optional: The package
API does not assume config files being passed around.
Rather, the vignette showcases how the .yml file is
parsed.

The config files will be more useful when a cli is
defined for the package, or to share project configs
around. Note the files are not fully reusable as it is,
because the end-user still has to specify the path for
any external dependencies.

Nonetheless, I expect as more functionality is added,
other parameters will still justify keeping the config
files. Such as Depends need to specify a language,
bug density metrics needing to specify commit hashes,
file filtering based on extensions or certain regex,
or even identity matching approach for cli.

This commit also adds to parser functions a path
expansion, which allows for config files to include
relative paths using ~, e.g. ~/Desktop.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>