i #52 make configs more reusable
External tools are now specified on tools.yml. The
parameters of the tools are still project specific,
as we want to encourage reproducibility.

Vignettes and batch script were updated to reflect the change
in configs accordingly.

Data on configs is now assumed to be on a gitignored folder named
rawdata. Provided this convention is respected, config files should
be fully reusable for each project after the 3 line tools.yml is set
up in a different machine.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>