i #54 enable use of git log --grep --perl-regexp
For very large git log analysis like Chromium in
which the log of interest is a much smaller subset,
for example a subset of security bugs or CVEs, being
able to use --grep makes the analysis viable.

The parse_gitlog() function now uses the git log --grep
interface to only provide the log containing the issues
of interest, considerably reducing its size and make it
suitable for analysis in R.

parse_gitlog() now accepts a new parameter containing a
perl regex. This uses a different set of parameters,
only including the master branch instead of all branches.

If the extra perl_expr parameter is not used, the function
will have its prior behavior including all branches for
exploration.

To highlight the discussion of vulnerabilities, as the
data is not present in every project, a new vignette dedicated
for git log and vulnerabilities parsing was created.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>