i #60 adds identification of line types
A wrapper for universal ctags was added and
associated vignette. The wrapper simplifies the output
table by providing a csv table and header to the output
of universal ctags.

Setup was added to README.md and a minimal vignette was
added showcasing the wrapper function call interaction
to existing config files.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>