i #38 adds testthat and identity tests
Add several tests for different formats of name in
mailing list and git. All tests names and emails have
been modified to preserve the test case but also respect
the privacy of each person, including the domain of the
email if not a generic domain like gmail.
Current identity fails several
of these tests, and an improved identity function which
pass all tests will be send on a follow commit.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>