i #68 add git blame parser
Adds a git blame parser. Note this is a git
command, and the filepath needs only to exist
in the git log to be output. Said filepath and
commit hash can be obtained, for example, using
parse_gitlog(). A vignette will be added in
a later commit showing how to combine both.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>