i #82 add filter for entities per time window
A new filter function in R/filter was added:
parse_by_last_files_change(). This function has
the equivalent behavior of git_checkout() and
searching for the folders for the existing files,
but leverages the already loaded in memory
parse_gitlog() output to reconstruct them.

This function can be used in combination to
parse_gitlog_entity() to analyze larger projects
such as OpenSSL, where blaming every commit can
take days.

The assumption of the function is that, within a
specified time window of interest, collaboration
is quasi-stationary and while a threat to validity,
can still be used and has been used in the past
in publications.

See vignettes/gitlog_entity_showcase.Rmd for usage.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>