Adds end-to-end issue assignment data pipeline
This commit integrates the necessary functionality
to obtain the list of issue assignments of a chosen
project. User only needs to specify the organization,
project name and folder to save raw data. Parsed data
and explanation of the tables and API is provided in
the R Notebook GitHub API Showcase in plain english.

A new API function was added to iterate over any
endpoint response and end execution without throwing
error.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>