i #75 Adds built-in bipartite projection
Kaiaulu now has a built-in bipartite projection.
The bipartite projection function allows to compute
two types of projections, which is reflected on the weight.

In one mode, which is equivalent to the method commonly used
in graph libraries such as igraph, the number of unique nodes
deleted is calculated as the projection weight.

In the new mode, the original weights in the bipartite graph
are sum across the deleted nodes. By default, Kaiaulu implements
this method.

It is also possible to return a intermediate projection table,
which shows the original weights and the name of the removed node,
in order to calculate the graph weights as done in other libraries
or other method for combining the original weights.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>