i #126 fixes smells txt to int maps and comm hash
Due to incorrect and incomplete txt to int maps
on the new smell re-implementation, the smells
were incorrectly being computed. This was observed
by a side-effect where the ordering of the table
influenced the metric value.

In the associated social smell notebook, the sorting
was also relocated so it is now properly done based
on time instead of alphabetically.

The first and last commit hash of each time slice
was also incorrectly being obtained, leading to
incorrect interval being shown in the last table
of the social smell notebook, and in turn passed
incorrectly to the line metrics (thus compromising
the line metric values). This is also fixed in this
commit.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>