i #134 fixes radio_silence with multi-groups
When multiple groups were assigned to a single
developer, the radio_silence function would
throw warnings for not knowing which group to choose
(as the default implementation only assigned one
community). For consistency, radio_silence now
chooses the first group (as did the warning).

In the future, the smell can be revisited to
account for multiple group assignment.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>