i #132 smells notebook now detect reply source
The social smells notebook will now use the project
configuration file to decide which data sources to
parse. If a communication channel path is not specified,
the notebook will skip the source parsing accordingly.

Some of the smell notebook narrative was also updated
with additional advice and threats to validity for
both JIRA and GitHuB comments data.

All variables named after 'mbox' in the code should
also now properly be referred as "reply" instead.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>