i #132 Add CLI to compute smells on branches
A smells interface was added to CLI to compute
smells over multiple branches. The code is in
essence the smells notebook within a loop of
loaded branches from the project configuration
file.

A single table will be generated containing the
smells analysis, in addition to a column specifying
the branch it is associated to.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>