i #185 parse_mbox only renames existing fields
parse_mbox() has been modified so it does not assume
every .mbox file constains a fixed number of fields.
Rather, it will only identify from a list of fields
of interest those that exist in the data, and rename
them accordingly.

The function documentation was also updated to indicate
due to this change, the number of columns in the resulting
table may change depending on the data.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>