i #184 fixes no changes file renaming
In instances when a commit modifies a file name
without changes, Perceval includes the new file
information under "newfile" field, as opposed of
"file".

For cases where after a file rename without content
changes, said file is never modified again in the
git log history, the new file name will never be
included in the parsed git log.

This fix now parses the "newfile" field and it
includes as part of the parse_gitlog() information.

It also overwrites the "file" field with the "newfile"
field information, so the new file name is correctly
included in the git log in cases where no future changes
happened to the file.

In the future, the "newfield" information can be used to
properly rename all past file ocurrences, for analysis
which treating the new file as a separated file could
be wrong.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>