i #192 removes notebook Suggests
Various suggests in Kaiaulu are mainly used in
Notebooks, and are not required to use the API.

Moreover, most of them are localized to only a few
notebooks. The dependency requirement should be
passed on as need basis to the notebooks that require
them instead.

pkgdown was removed due to errors on GitHub actions.
Moreover, currently there is no way to actually use
pkgdown in other machine unless the data it relies
on for showcase is also included. In the future,
as Kaiaulu matures, notebooks should leverage
Kaiaulu own logs to present features.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>