i #192 adds testthat output
While the R CMD check workflow actually runs
testthat, when it is successful it only outputs
"OK". This includes on test/coverage the full
output of tests for inspection if needed.

If a unit test fail, both R CMD and the code coverage
will both fail, and provide the full unit test output
too.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>