i #184 modularizes bipartite projections
bipartite_graph_projection() now accepts a weight_scheme
function to decide how to assign weights to the nodes of
the resulting projection. When a function is missing,
the intermediate projection table is returned instead.
This can be useful for users trying to create their own
weight function to the API, or just to utilize as-is.

Commonly, projections are performed as weight sum. A new
weight scheme function, weight_scheme_count_deleted_nodes(),
can now also be used to assign weight to the projection edges.
This function is appropriate, for example, when computing the
co-change metric.

Various notebooks which relied on the function, and other API
functions were also updated accordingly with the appropriate
weight schemes.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>