i #184 adds kaiaulu dsmj to DV8 Notebook
Kaiaulu's gitlog_to_hdsmj() and
dependencies_to_sdsmj() are now used and
combined to filters in the DV8 Showcase Notebook.

An additional conf file to a simpler project to serve
as example was also added, given the flaws mapping to
APR cost extensive computational time when generating
the documentation.

For the calculator project, however, the excel export
appears to generate a corrupted file, and code
evaluation was removed for the Notebook until further
testing is done.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>