i #153 Kaiaulu's Bugzilla Downloader (#177)
Adds downloader functions for Kaiaulu to download
Bugzilla JSON for issues and comments separately. 

These functions do not depend on Perceval. Moreover,
the downloader provides files "as it runs", so the task 
does not need a full re-start in case of failure.

Guesswork on the API limit is also removed here, as the 
function will use the number of downloaded files per page
to infer the subsequent steps.

---------

Signed-off-by: Leilani Reich <reichld@hawaii.edu>
Co-authored-by: malialiu <mliu0141@gmail.com>
Co-authored-by: Carlos Paradis <carlosviansi@gmail.com>