i #200 fixes check and tests, and actions
When executing the commands devtools:check() and
devtools::test(), files specified in unit tests can
be located by devtools::test() by assuming the function
is executing from kaiaulu/tests/testthat ; however this
assumption is not valid for devtools::check()

The end result is that the checks always fail, and thus
the GitHub actions will also always fail.

The solution was found on https://stackoverflow.com/a/71334780
which specifies files which unit tests rely on should be placed
within the unit test folder. Moreover, said file paths should be
referred by a special function test_path().

When defined in this manner, both the check and test function will
work, as the file paths will be adjusted accordingly. This
unfortunately introduces a redundancy in the files, and also
partially defeats the intent of some of these tests: I.e. to
aid the user in verifying if their tools.yml define the correct
path. Future README.md recommendations to run unit tests will
hence require the user to copy their tools.yml to the testdata
folder to execute said assumption.

This commit also fixes GitHub Actions to install perceval before
executing the tests. Moreover, the Actions will update the
tools.yml with the remote machine path to the installed perceval.
Due to what is described above, GitHub Actions will also copy on
the remote machine only the updated tools.yml to the testdata folder,
as someone would need to do manually, so the unit tests can verify
the Perceval setup was done properly.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>