i #202 #203 Update DV8 Notebook and Bugzilla API
 * Modifies Bugzilla API to allow for output files
 * Updates DV8 Notebook to use more general functions

The Bugzilla showcase is currently not available via
the docs compilation, but is executable locally and can
be knitted. An issue occurs during the documentation
process involving nulls which requires further verification.

This likely occurs due to the text content included on issues
causing some internal conflict with pkgdown html generation.
Since code is functional and reusable, the modifications
are being versioned.

This commit also moves Kaiaulu to version 0.0.0.9700.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>