i #209 refactor graph_to_dsmj
The currently performance issue lies on
getCells. It iterates through every edge pair
and converts the "weight/label" into a data.frame.

This loop runs fine on 140k or so rows, which
seems typical for the SDSM, however, for HDSM
which requires projecting the gitlog to obtain
co-change, edgelists can increase to over 7 million,
which makes the code runs eternally.

the reverse getCells then loops through the 7 million
list again.

This implementation removes the reverse Cells and the
loop and vectorize some parts, but still suffers from
the same issue.

Perhaps the solution lies in performing the operation
during the projection operation itself.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>