i #206 #211 Adds GoF/Text Parsers
This adds two methods to the identification of GoF 
Patterns, and a text module to Kaiaulu. 

The text module is a wrapper to srcML.
`text.R` defines a basic interface to obtain
the XML and a sample function to query it 
and tabulate the output to R memory. 

The GoF pattern relies on pattern4.jar by
Tsantalis. It wraps around the jar and 
tabulates the output to R memory. 

An example of how to use both new third
party tools can be found on gof_showcase.Rmd

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>