i #222 Parameterize dsm types for Flaws
When calculating architectural flaws, it is not
always the case a project will have both HDSM and
SDSMs available (and therefore MDSMs). This
parameterizes the function call so the user can
specify which of the 3 types they prefer the
function to use when calculating the file instances
of architectural flaws.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>