i #244 add issue fake example and metric tests
A new module, jira.R was added with just a simple
example of fake data for replicating issues data
that allowed the identification of bugs in Kaiaulu
and the asssociated unit tests to avoid them in
the future. The paired fake example data and unit test
currently fail. A subsequent commit will address the
issue allowing the test to pass.

A few unit tests were also added to the metric module,
which was where the side effect of the parse_jira bug
was first reported.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>