i #244 fixes jira log duplicated issues
The presence of multiple components in a jira id caused
the resulting issue table to be duplicated. The unit test
which captured that bug now passes. Moreover, the churn
metric value reported in the issue which had as a side
effect a incorrect file_non_bug_churn metric also now
reports the correct value.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>