i #245 #246 move Kaiaulu to 4.0 and dv8 f-metrics
Move Kaiaulu to version 4.0. Minor changes to a few
notebooks in the process. The bump is due to XML
now only being available on version 4.0, which causes
random errors on GitHub Actions if the version
requirement is under version 4.

Also added a new parameter to dv8_mdsmb_to_flaws,
is_file_only_metric, which leverages the built-in
computation of file metrics in DV8. This is substantially
faster than generating the file to flaw map, but can't be used
for aggregation (or files that participate in the same flaw
id will be counted twice when aggregating).

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>