i #229 Add general weighted temporal graph
Currently, there exists in network.R a simple
implementation of the temporal graph projection.
The transform creates a graph with model_direct_graph,
and then peforms the temporal transformation.

This commit adds a more generalized temporal transform,
which is consistent to the new version of graph_projection.
The code is currently on branch because the model_direct_graph
had to be bent somewhat to maintain its old interface,
while allowing for this one to exist, which currently
bypass a transform function from network.R.

More details on the issue #229.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>