Adds cumulative temporal projection
This implementation is the actual equivalent
to the original formulation defined in Joblin.

Given a contribution of Dev C, after Dev A and Dev B
commits in this order, then Dev C -> Dev A, and Dev C -> Dev B.

In the non cumulative, Dev C -> Dev B (no links between C and A).

I have also inverted the direction of the edges so it is consistent
to the original formulation, and more consistent to overall notion
of "Depends". E.g. if Dev C -> Dev A, it is because Dev C had to
read code from Dev A. Originally, the formulation was Dev A -> Dev C
in which case the edge reflected the flow of time.

The code and variable naming is still very poor, and needs more
refactoring before going to master. But it is a functional example.

There are two unit tests here: One validates the results match the
original dissertation example. The other tests a more complex example,
where the developers alternate turns on changing the same file. This
means there will be collaboration edges flowing on both directions
between the same developers, however they will not have the same
weight.

See issue for more details.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>