i #229 Refactor Transforms for Temporal
- Fixed a bug when a file only had a single
developer changing the same file.

- The temporal transformations now properly use
the new graph temporal projections, and are
parameterized to specify the weight scheme
and the lag. I believe trying to use the cumulative
sum intended for all_lag on other networks may break
but there are warnings on the function description :^)

- All unit tests now use the transforms and pass.

- Both the Gitlog and Git Entity notebooks now use the
transform. I also changed their proj to Kaiaulu, as
the log is now reasonable large enough to be able
to see the functions in action and reason. I've used
the git log on one lag to showcase, and the git entity
on all lag to showcase.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>