i #227 #108 more fake git data and fixes
- More unit tests and example fake datasets were added

- One of the unit tests, that is to check empty git repo, 
actually relates to #108 and was further addressed.

Specifically, when using parse_gitlog() to parse an empty
repo (but existing repo, and therefore with 0 commits), 
calling the `git log` command will throw a fatal error, 
since it can't generate logs for 0 commits. 

The system call interprets that as a warning, returning 
a NULL object, instead of halting execution. Eventually, 
in the parse_gitlog() function, Perceval is called, likely
throwing another warning but also failing to halt 
execution. 

At some point, a transformation on the naming of the columns
will occur, including "data.Author" which is what Perceval calls
the author_name_email column. By then, it is data.table who throws
an error saying the column doesn't exist, leading to the problem 
experienced, which makes no sense to anyone reading. 

This commit also patches this problem, and Kaiaulu will now capture 
empty git logs errors and fail gracefully properly warning the user
either they pointed to a folder that doesn't exist, or that it lacks 
any commits.

---------

Signed-off-by: Waylon Ho <84744515+waylonho@users.noreply.github.com>
Co-authored-by: Carlos Paradis <carlosviansi@gmail.com>
Co-authored-by: Ruben Jacobo <rjacobo@hawaii.edu>