Added file iteration and ability to return latest date
I have added the functionality to parse_jira function that allows it to go through each of the json files in the save directory to parse them all. I also made it so that it finds the latest created date of an issue and returns it so that Sean may get the needed info for use in the refresher; it is in YYYY-MM-DD format. As far as I'm aware, Sean DID NOT change the config file at all for what the save_path is, so the variable "save_path_issue_tracker_issues" and "save_path_issue_tracker_issue_comments" should still be pointing at a single json file. Due to this, I have made chosen accordingly on how I can determine which file patterns the parse_jira function should look for. If it isn't satisfactory with the way I did it, I can easily change it.