Changed file iterator logic, and method finding latest date
I changed the way the parser function retrieves and iterates through the json files. Also, I changed the way that it searches the latest issue date: I made it so that it takes the upper bound unix date in the file names and finds the max among those. Also, I tried building documents and buildin kaiaulu but it still does not pass checks, I may need a bit of help, or maybe it worked I'm not sure.