i #275 Modified refresh to use 'issueKey'
download_and_save_jira_issues_refresh() has been added. It takes in the normal parameters (-search_query) but also a filename. The intention is that this filename will contain the greatest issueKey value of all files in a path. The function extracts this value and appends it to the API query so only issues greater than this value are downloaded.