i #241 Refactor Kaiaulu Architcture
The overall architecture of Kaiaulu has been refactored
from "downloaders/parsers/networks" to an architecture
that reflect the data source or tool interface, i.e.:
source code/git/mail/jira/bugzilla/ etc. The functionality
of the downloaders, parsers and network transformations
are then now in the respective modules.

This reflects the intuition that a user will seek for, for example,
Git functionality, or want to download JIRA issues, instead of
broadly look to browse all parsers in Kaiaulu. A benefit
of this new architecture is that the parsers, downloaders
and network functionality now will no longer grow in size
to every new tool or data source added, nor the associated
unit test files.

For more details and rationale, see the respective issue for
this commit.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>