Added missing fields in parser, fixed documentation
I added the missing extra fields for the parser to parse. Along with that, I fixed the documentation for the two parser functions for JIRA. From what I could tell based on other function documentation, I put the note in the wrong spot (i.e. after parameters), so I moved them so that they're before the parameters are documented.