i #282 added Check to Refresher
Added a check to both issue and comment refresher that checks if the directory is empty before running. If the directory is empty, it will run the regular downloader because otherwise it will break when trying to parse file names. issue refresh now downloads raw data and does not alter it but github_iterate pages cannot write it to file yet because it depends on an altered format