i #275 Project key is now from JQL, and new Creds
- Previously, project key was obtained from the config
issues path as reflected in an additional sub-folder
under issues. The key is now obtained from the JIRA
issue key (specifically the portion project='PROJECTKEY'.

- Minor grammar fixes, and updates to the notebook to
reflect that were added.

- Credentials have been broken down into username and
password parameters instead of expecting a file path.
The file is still expected, but it is handled by the
notebook instead.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>