i #275 Fixes authentication requests
Authentications requests did not work on Kaiaulu
JIRA. This commit fixes the authentication issue.

Data can be downloaded both with and without
authentication (tested on geronimo w/o and on
kaiaulu with).

In addition, httr verbose was incorporated, which
helps with troubleshooting. The agent now also
is specified, pointing to Kaiaulu repo.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>