i #275 remove unaltered_file_path and fix auth
Reorganized the condition surrounding this file.
The logic was already using it as a saved_folder_path.

Renamed the `contents` variable to something more clear.

A lot of the code surrounding auth was also fixed and
expanded on.

At the stage of this commit, the downloaders work
for date range and issue range, and the refresher
also works (the refresher was tested on Kaiaulu
JIRA where I have more control over the timestamps).

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>