i #275 added extra folder refresh_issues for issue refresh
I also fixed the naming convention for the commit data. Fixed some verbose wrappers around messages. Added parameter to format_created_at_from_file to specify nesting levels. If issue folder is empty, it will download all issues. This is also the case if the issue folder is empty but there is data in the refresh folder. Documentation was not worked on thoroughly, but minor adjustments were made. Also, an error is output when trying to refresh the comments if there is nothing to refresh. My other refresher outputs a formal message for this, but I'm not sure what's causing it for the comment refresher.