i #302 Changes empty data frame header
During entity analysis, we save an empty data frame in case no entities
were found in the respective time window. This indicates that a specific
range has not been skipped accidentally, but did not contain any changed
entities. Change the header of this data frame match the standard format
to facilitate subsequent analyses.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>