i #302 Refactors file and entity options for git CLI
In the project configuration file, we can specify options such as file
filters which can be applied to file-based or entity-based analysis
modes or both.

So far, the application of these options was hard-coded in both git
CLIs.

Now, users may specify the desired options and their application to
file-based and entity-based analysis modes separately in the CLI
configuration file. This gives users more flexibility in their analyses.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>