i #312 update notebooks for jira downloader
The new JIRA downloader removed the dependency on JirAgileR. However, some notebooks still depended on it, which made compiling the documentation impossible. This fixes that. The commit also fixes a new issue with GitHub Actions by bumping the Actions (not Kaiaulu language requirement) to 4.4. The XML dependency issue should not happen local regardless of the R version.

The pkgdown API for JIRA was also updated. This commit docs can be generated.

---------

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>