i #230 Incorrect assumption of .git placement
Originally, it was assumed that .git could reside
folder-less to be accessed. While this is true, it
adds inconvenience to the user since when data is obtained
with .git, it is gererally through git clone, which
also adds the source code. The consequential effect is that
the path will be project/git_repo/project/.git rather than
project/git_repo/.git.

Not all configs were updated to adjust for this, only the
ones used in the notebooks.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>