Documentation pass
Simplified some of the notebook language, reduced
title of functions, removed some of the sub-headers
pound symbols as it was creating too many sections
on the code blocks.

Added parser tables after downloaders and remove their
eval so example tables of what can be downloaded are
shown on the generated notebook.

Commit passes check, tests, and downloaders, refresh
and parsers work.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>