Makes raw .xml files to be saved on folder
This modifies the XML files so they are saved locally.
This adds an additional function step, removes the scitools
and dependency from parser.

The independent parser function means the raw data is not locked
behind the db, should a new version make it incompatible to extract
in the future. It also facilitates inspection of the raw data
should the parser ever break in the future.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>