i #323 Replaces PR Refresher with Wrapper
In the Notebook, the github_api_project_pull_request_refresh()
was not utilized as the endpoint it relied on did not accept a timestamp.
Instead, an existing function was used in Kaiaulu as a workaround.

As such, the github_api_project_pull_request_refresh() was modified
to simply serve as a wrapper around said function. In this manner,
the notebook relies to the appropriate function on the cheatsheet and
the functionality is preserved.

Signed-off-by: Carlos Paradis <carlosviansi@gmail.com>